Added 'Max Cap Timeout' for help channels
Added a feature where a channel cannot stay open past a specified amount of time. The default value for cap is 3 days, but may be changed by configuration.
The command for configuration is '.config clopen <server id> timeout_cap [duration]'

This is tested locally with different amount of times ranging from a few minutes to an hour. I did not test for larger values since I was impatient. But it should still work.

More specifications:
Upon closing, a reason for closing will be given "Channel closed due to maximum timeout reached!".
The user is not however pinged as this is an embedded message.
When closed, the channel will be archived (moved to hidden help channel category in the server).