Refactored code and fixed condition checking
Refactored the code for closing channel due to maximum timeout reached.
Condensed the whole procedure of closing a channel due to max timeout reached into a function called `timeout_cap_close_procedure`

Fixed condition checking:
Fixed the condition checking in `scheduler_task`. The structure should be if-elif instead of if-if when checking conditions for max_expiry and expiry.